Restructure and small bits of styling
